remove auth1.service.ts and fix component influenced from that
