+++ src/app/app.component.ts
@@ -1,4 +1,4 @@
-import { Auth1Service } from './dev/services/auth1.service';
+import { AuthService } from './dev/services/auth.service';
 import { Component, OnInit } from '@angular/core';
 
 @Component({
@@ -12,7 +12,7 @@ export class AppComponent implements OnInit {
   title = 'app';
 
   constructor(
-    private auth: Auth1Service,
+    private auth: AuthService,
   ) {}
 
   ngOnInit() {
