+++ src/app/dev/services/auth.service.ts
@@ -1,7 +1,7 @@
-import { Injectable } from '@angular/core';
-import { HttpClient } from '@angular/common/http';
+import { Injectable, Inject } from '@angular/core';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { Observable } from 'rxjs/Observable';
-import { environment } from './../../../environments/environment';
+import { environment } from '../../../environments/environment';
 
 import 'rxjs/add/observable/of';
 import 'rxjs/add/operator/do';
@@ -24,19 +24,30 @@ class TryLoginUser {
   constructor(
     public email: string,
     public password: string
-  ) {}
+  ) { }
 }
-
 
 @Injectable()
 export class AuthService {
   appUrl = environment.apiUrl;
-  TOKEN_NAME = 'jwt_token';
-  constructor(private http: HttpClient) { 
+  TOKEN_NAME = 'token';
+  PK_NAME = 'user_pk';
+
+  constructor(private http: HttpClient) {
     console.log('[appUrl] ', this.appUrl);
   }
-  signin(user: TryLoginUser): Observable<SuccessLoginUser> {
-    return this.http.post<SuccessLoginUser>(`${this.appUrl}/auth/login/`, user)
+
+  joinIn(signupForm) {
+    return this.http.post(`${this.appUrl}/auth/signup/`, signupForm)
+      .do(res => {
+        console.log(res);
+        console.log('회원가입 성공!');
+      });
+  }
+
+  // 로그인
+  login(loginForm: TryLoginUser): Observable<SuccessLoginUser> {
+    return this.http.post<SuccessLoginUser>(`${this.appUrl}/auth/login/`, loginForm)
       .do(res => {
         this.setToken(res.token);
         this.setUserPk(res.user.pk);
@@ -44,23 +55,42 @@ export class AuthService {
       .shareReplay();
   }
 
-  setToken(userToken: string): void {
-    localStorage.setItem(this.TOKEN_NAME, userToken);
+  // 로그아웃
+  logout() {
+    let headers = new HttpHeaders();
+    headers = headers.append('Authorization', `Token ${this.getToken()}`);
+    return this.http.post(`${this.appUrl}/auth/logout/`, null , { headers: headers })
+      .subscribe(res =>
+        this.removeTokenAndPk(res)
+      );
+  }
+
+  // 토큰 유효성 검증
+  isAuthenticated(): boolean {
+    const token = this.getToken();
+    return token ? true : false;
+  }
+
+  setToken(token: string): void {
+    localStorage.setItem(this.TOKEN_NAME, token);
   }
 
   getToken(): string {
     return localStorage.getItem(this.TOKEN_NAME);
   }
 
-  setUserPk(userPk): void {
-    localStorage.setItem('loginUser_pk', userPk);
-  }
-  getUserPk() {
-    return localStorage.getItem('loginUser_pk');
-  }
-  removeTokenAndPk(): void {
-    localStorage.removeItem(this.TOKEN_NAME);
-    localStorage.removeItem('loginUser_pk');
+  setUserPk(pk: any): void {
+    localStorage.setItem(this.PK_NAME, pk);
   }
 
+  getUserPk(): string {
+    return localStorage.getItem(this.PK_NAME);
+  }
+
+  removeTokenAndPk(res): void {
+    console.log(res.message);
+    localStorage.removeItem(this.TOKEN_NAME);
+    localStorage.removeItem(this.PK_NAME);
+    console.log(localStorage);
+  }
 }
