--- src/app/dev/user/pet-register-test/pet-register-test.component.html
@@ -1,29 +1,0 @@
-
-<div class="container">
- 
-  <!--<div class="pet-profile">
-    <h2 class="pet-profile-pic">
-      <img src="" alt="pet"/>
-   </h2>
-    <p class="pet-name">pet-name</p>
-    <div class="box">
-        <button class="replace"></button>
-        <input type="file" value="파일 업로드" class="upload">
-    </div>  
-  </div>
--->
-
-  <div class="pet-profile-info">
-      <!--form-->
-     
-        <form [formGroup]="petForm">
-        
-        </form>
-      <!--form-->
-      <button (click)="test()">test</button>
-  </div>
-  <!--pet-container-->
- <div>
-    {{ petForm.value | json }}
-    {{ petForm.status }}
-  </div>
